Changes were made to refactor the code and improve the color display in the character display module. The changes include:
