+++ __init__.py
@@ -1,0 +1,2 @@
+# dnd_adventure/__init__.py
+# Empty file to make dnd_adventure a package
