+++ dnd_adventure/race_selector.py
@@ -11,10 +11,10 @@ def select_race(races: List) -> str:
         print(f"{Fore.CYAN}=== Select Your Race ==={Style.RESET_ALL}")
         print(f"{Fore.LIGHTBLACK_EX}----------------------------------------{Style.RESET_ALL}")
         for i, race in enumerate(races, 1):
-            desc = race.description[:100] + "..." if len(race.description) > 100 else race.description
+            desc = race["description"][:100] + "..." if len(race["description"]) > 100 else race["description"]
             modifiers = getattr(race, 'stat_bonuses', getattr(race, 'ability_modifiers', {}))
             modifiers_str = ", ".join(f"{k}: {v:+d}" for k, v in modifiers.items()) or "No modifiers"
-            print(f"{Fore.YELLOW}{i}. {race.name}{Style.RESET_ALL}")
+            print(f"{Fore.YELLOW}{i}. {race["name"]}{Style.RESET_ALL}")
             print(f"     {Fore.LIGHTYELLOW_EX}{desc}{Style.RESET_ALL}")
             print(f"     {Fore.LIGHTYELLOW_EX}Modifiers: {modifiers_str}{Style.RESET_ALL}")
             print(f"{Fore.LIGHTBLACK_EX}----------------------------------------{Style.RESET_ALL}")
@@ -40,11 +40,11 @@ def select_race(races: List) -> str:
 def select_subrace(subrace_names: List[str], race) -> str:
     while True:
         os.system('cls' if os.name == 'nt' else 'clear')
-        print(f"{Fore.CYAN}=== Select Subrace for {race.name} ==={Style.RESET_ALL}")
+        print(f"{Fore.CYAN}=== Select Subrace for {race["name"]} ==={Style.RESET_ALL}")
         print(f"{Fore.LIGHTBLACK_EX}----------------------------------------{Style.RESET_ALL}")
         for i, subrace in enumerate(subrace_names, 1):
             if subrace.startswith("Base "):
-                desc = f"Standard {race.name} with no subrace-specific traits."
+                desc = f"Standard {race["name"]} with no subrace-specific traits."
                 modifiers = "No additional modifiers"
             else:
                 desc = race.subraces[subrace]["description"]
